Make emoticons smoother

+++ src/game/client/gameclient.h
@@ -257,7 +257,7 @@ public:
 		int m_SkinColor;
 		int m_Team;
 		int m_Emoticon;
-		int m_EmoticonStart;
+		float m_EmoticonStart;
 		bool m_Solo;
 		bool m_Jetpack;
 		bool m_NoCollision;
